Major refactor. All views and models moved into their own files.
The 'accept license' modal is actually a standalone modal. It
used to be an installer pane before.

The subviews for the individual installer steps will be recreated
if the user navigates back and forth repeatedly. This will leak
memory.
