+++ macInstaller/macInstaller/Models/InstallSummary.swift
@@ -1,0 +1,13 @@
+//
+//  InstallSummary.swift
+//  macInstaller
+//
+//  Created by Paulo Raffaelli on 1/23/25.
+//
+
+enum InstallSummary {
+    case licenseRejected
+    case installationSucceeded
+    case failed
+    case failedWithRollback
+}
